requirements.txt
livekit-agents[silero,openai,elevenlabs,deepgram]
python-dotenv
